pre final pharse 1
