+++ Drive sound/web project/About_Us_Page.html
@@ -29,7 +29,7 @@
                                 <p>6388001 Phumtham Akkarasiriwong</p>
                                 <p>Section 1 Group 6</p>
                                 <h4>Contact</h4>
-                                <p>phumtham.akk@student.mahidol.ac.th</p>
+                                <p><a href="mailto: phumtham.akk@student.mahidol.ac.th">phumtham.akk@student.mahidol.ac.th</a></p>
                             </div>
                     </div>
                 </div>
@@ -44,7 +44,7 @@
                                 <p>6388048 Leenawat Honglerdnapakul</p>
                                 <p>Section 1 Group 6</p>
                                 <h4>Contact</h4>
-                                <p>leenawat.hon@student.mahidol.ac.th</p>
+                                <p><a href="mailto: leenawat.hon@student.mahidol.ac.th">leenawat.hon@student.mahidol.ac.th</a></p>
                             </div>
                     </div>
                 </div>
@@ -59,7 +59,7 @@
                                 <p>6388137 Phurich Kongmont</p>
                                 <p>Section 1 Group 6</p>
                                 <h4>Contact</h4>
-                                <p>phurich.kon@student.mahidol.ac.th</p>
+                                <p><a href="mailto:phurich.kon@student.mahidol.ac.th">phurich.kon@student.mahidol.ac.th</a></p>
                             </div>
                     </div>
                 </div>
