+++ Drive sound/web project/HomePage.html
@@ -20,7 +20,10 @@
                 <div class ="bg-image">
                     <p class="title">Welcome back user</p>
                 </div>
+                
                 <div class ="container">
+
+                    
                     <a href="HomePage.html" style="text-decoration: none"><h1> New in drive sound</h1></a>
                     <div class="row">
                         <div class="column">
