+++ Drive sound/web project/Login.html
@@ -1,58 +1,5 @@
 <!DOCTYPE html>
 <html>
-<style>
-    * {
-        font-family: 'Noto Sans KR', sans-serif;
-    }
-
-    body {
-        font-family: Arial, Helvetica, sans-serif;
-        margin: 0;
-        background: rgb(226, 234, 255);
-    }
-    
-    #im {
-        float: left;
-        padding: 14px 16px;
-        padding-top: 5px;
-        padding-bottom: 5px;
-        text-align: center;
-    }
-
-    .container {
-            margin: auto;
-            width: 500px;
-            max-width: 90%;
-            align-items: center;
-            display: flex;
-        }
-
-        .container form {
-            width: 100%;
-            height: 100%;
-            padding: 20%;
-            background: white;
-            border-radius: 4px;
-            box-shadow: 0px 18px 16px rgba(0, 0, 0, .3);
-          
-        }
-
-        .container form h1 {
-            text-align: center;
-            margin-bottom: 24px;
-            color: #222;
-        }
-
-        .container form .form-control {
-            width: 100%;
-            height: 40px;
-            background: white;
-            border-radius: 4px;
-            border: 1 solid silver;
-            margin: 10px 0px 18px 0px;
-            padding: 0 10px;
-        }
-</style>
 <head>
     <title>Login page</title>
     <link rel="stylesheet" href="css/styleforlogin.css">
