+++ Drive sound/web project/Result_Page.html
@@ -1,56 +1,113 @@
 <!DOCTYPE html>
 <html>
-
 <head>
     <title>Result Page</title>
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="stylesheet" href="css/Option.css">
-</head>
+    <link rel="stylesheet" href="css/styleforresult.css">
 
+</head>
 
 <body>
     <header class ="header">
-        <a href="HomePage.html"><img src="images/Logo_DriveSound-01.png" width="100px%"></img>
-        <nav class ="nav">
-            <a href="Search_Page.html">Search</a>
-            <a href="About_Us_Page.html">About Us</a>
-            <a href="Account.html">Account</a>
-        </nav>
+        <a href="HomePage.html"><img src="css/images/Logo_DriveSound-01.png" width="100px%"></img></a>
+    <nav class ="nav">
+        <a href="Search_Page.html">Search</a>
+        <a href="About_Us_Page.html">About Us</a>
+        <a href="Account.html">Account</a>
+    </nav>
     </header>
-    <div class="header">
-        <p>Result</p>
-        <br>
-        <br>
-        <br>
-        <br>
-    <div class="container">
-
-        <table style="display: none;" id="table" class="table table-striped">
-            <thead>
-                <tr>
-                    <th style="background-color:white;">Title </th>
-                    <th style="background-color:white;">Cover Photo</th>
-                    <th style="background-color:white;">Channel</th>
-                </tr>
-
-            </thead>
-            <tbody id="results">
-            </tbody>
-        </table>
-        <div id="pager">
-            <ul id="pagination" class="pagination-sm"></ul>
+    <div class ="container">
+        <div class="row">
+            <div class="column-left">
+                <div class="card">
+                    <div class ="img-cover">
+                        <br>
+                        <img src="css/images/result/sugar.jpg" width="100%" height="auto">
+                    </div>
+                </div>
+            </div>
+            <div class="column-mid">
+                <div class="card">
+                    <h1>Sugar</h1>
+                    <hr>
+                    <p>Produced on xx xxx xxxx</p>
+                    <p>Rating : &#11088;&#11088;&#11088;&#11088; (140) </p>
+                    <p>Brand : <a href="Result_Page.html">Maroon Five</a></p>
+                    <hr>
+                    <div class = "col-mid-upper-nav">
+                        <nav>
+                            <br>
+                            <a href ="Result_Page.html" >Detail</a>
+                            <a href="Result_Page.html">Chord</a>
+                            <a href="Result_Page.html">Lyric</a>
+                            <a href="Result_Page.html">Album</a>
+                        </nav>
+                     </div>
+                     <div class ="detail">
+                         <h2>Detail</h2>
+                        <p>
+                            <br>Description of Sugar..............................................................
+                            <br><br><br><br><br><hr>
+                        </p>
+                     </div>
+                     <div class ="chord"></div>
+                     <div class ="lyric"></div>
+                     <div class ="album"></div>
+                </div>
+            </div>
+            <div class="column-right">
+                <div class="card">
+                    <li><a href ="Result_Page.html" >Listened</a></li><br>
+                    <li><a href ="Result_Page.html" >Like</a></li><br>
+                    <li><a href ="Result_Page.html" >Add to List</a></li>
+                </div>
+                <div class ="next-page">
+                    <a class="button"href="Result_Page.html" role="button" ><span class="text">Next Song</span></a>
+                </div>
+            </div>
         </div>
     </div>
     <footer class ="footer">
         <f>@drive sound</f>
     </footer>
-</body>
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/2.4.4/umd/popper.min.js"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js"></script>
-<script src="jquery.twbsPagination.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/js/bootstrap-datepicker.min.js"></script>
 
-<script src="script.js"></script>
+
+
+
+    <!-- <div class="header1">
+        <p id="i">Result</p>
+        <br>
+        <br>
+    <h1 style="text-align:center">SUGAR</h1>
+    <br>
+
+    <audio controls autoplay>
+        <source src="maroon-5-sugar.mp3" type="audio/mpeg">
+      </audio>
+    <br>
+    <br>
+
+    <div class="row">
+        <div class="column">
+            <div class="card">
+                <img src="sugar.jpg" style="width:60%" style="height: 50%">
+                <div class="container">
+                    <br>
+
+                    <h2 class="title" style="background-color:powderblue;">Maroon 5</h2> <br>
+                    <p style="text-align:left"> Maroon 5 is an American pop rock band from Los Angeles, California. It currently consists of lead vocalist Adam Levine, keyboardist and rhythm guitarist Jesse Carmichael, lead guitarist James Valentine, drummer Matt Flynn, keyboardist
+                        PJ Morton and multi-instrumentalist and bassist Sam Farrar. Original members Levine, Carmichael, bassist Mickey Madden, and drummer Ryan Dusick first came together as Kara's Flowers in 1994, while they were still in high school.</p>
+                    <p></p><br>
+                    <img src="Sugar2.png" style="width:60%" style="height: 50%"><br>
+                    <br>
+                    <h2 class="title" style="background-color:powderblue;">Sugar</h2>
+                    <p style="text-align:left">
+                        <br>"Sugar" is a song recorded by American band Maroon 5 for their fifth studio album V (2014). It was written by Mike Posner, Adam Levine, Dr. Luke, and Jacob Kasher Hindlin together with its producers Ammo and Cirkut. It was
+                        sent to contemporary hit radio in the United States, as the third single from the album on January 13, 2015. "Sugar" is a disco, funk-pop, and soul song that features a wide range of instruments including percussion, keyboards
+                        and guitars.
+                    </p>
+                </div>
+            </div>
+        </div> -->
+</body>
 
 </html>
