+++ Drive sound/web project/Search_Page.html
@@ -1,50 +1,64 @@
 <!DOCTYPE html>
 <html>
-
 <head>
-    <title>Search Page</title>
-    <link rel="stylesheet" href="css/styleforsearch.css">
+    <title>Search_Page</title>
+    <link rel="stylesheet" href="css/styleforsearch.css"> 
 </head>
+
 <body>
-    <header class ="header">
-        <a href="HomePage.html"><img src="css/images/Logo_DriveSound-01.png" width="100px%"></img>
-        <nav class ="nav">
-            <a href="Search_Page.html">Search</a>
-            <a href="About_Us_Page.html">About Us</a>
-            <a href="Account.html">Account</a>
-        </nav>
-    </header>
+<header class ="header">
+        <a href="HomePage.html"><img src="css/images/Logo_DriveSound-01.png" width="100px%"></img></a>
+    <nav class ="nav">
+        <a href="Search_Page.html">Search</a>
+        <a href="About_Us_Page.html">About Us</a>
+        <a href="Account.html">Account</a>
+    </nav>
+</header>
+<div class="headline">
+    <p>Search Box Music</p>
+</div>
     <div class="container">
-            <h2>Search Box Music</h2>
-        <br>
-        <br>
         <form id="myForm" autocomplete="off">
             <div class="searchBox">
-                <input class="searchInput" id="search" type="search" name="" placeholder="Search Music">
+                <input class="searchInput" id="search" type="search" name="" placeholder="Search Music" required>
                 <button class="searchButton" id="btn" onclick="SubmitFrom()">
-                   <i class="click" > 
+                   <i class="click" id="c" > 
                         Search 
                     </i>
                 </button>
             </div>
         </form>
     </div>
-    <footer class ="footer">
-        <f>@drive sound</f>
-    </footer>
+
+    <div class ="container1">
+        <table style="display: none;" id="table" class="table table-striped">
+            <thead>
+                <tr>
+                    <th style="background-color:#5d6475;" style="text-align:center;">Title</th>
+                    <th style="background-color:#5d6475;" style="text-align:center;">Cover Photo</th>
+                    <th style="background-color:#5d6475;" style="text-align:center;">Channel</th>
+                </tr>
+            </thead>
+            <tbody id="results">
+            </tbody>
+        </table>
+        <div id="pager">
+            <ul id="pagination" class="pagination-sm"></ul>
+        </div>
+    </div>
 </body>
 
 <br>
 <br>
 
 
-<script type="text/javascript">
-    function SubmitFrom() {
-        var Searchtxt = document.getElementById("search").value;
-        console.log("Searchtxt URL: " + Searchtxt);
-        // window.location = "Result_Page.html?Query=" + Searchtxt;
-        window.open("Result_Page.html?Query=" + Searchtxt);
-    }
-</script>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/2.4.4/umd/popper.min.js"></script>
+<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js"></script>
+<script src="jquery.twbsPagination.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/js/bootstrap-datepicker.min.js"></script>
+
+<script src="script.js"></script>
+
 
 </html>
