+++ Drive sound/web project/Search_Page1.html
@@ -1,0 +1,71 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Search_Page</title>
+    <link rel="stylesheet" href="css/styleforsearch.css"> 
+</head>
+
+<body>
+
+    <header class ="header">
+            <a href="HomePage.html"><img src="css/images/Logo_DriveSound-01.png" width="100px%"></img></a>
+        <nav class ="nav">
+            <a href="Search_Page.html">Search</a>
+            <a href="About_Us_Page.html">About Us</a>
+            <a href="Account.html">Account</a>
+        </nav>
+    </header>
+
+    <div class="headline">
+        <p>Search Box Music</p>
+    </div>
+    <div class="container">
+        <form class="myForm" autocomplete="off">
+            <div class="searchBox">
+                <input class="searchInput" id="search" type="search" name="" placeholder="Search Music" required>
+                <button class="searchButton" id="btn" onclick="SubmitFrom()">
+                    <i class="click" id="c" > 
+                         Search 
+                    </i>
+                </button>
+            </div>
+        </form>
+    
+
+        
+        <table style="display: none;" id="table" class="table table-striped">
+            <thead>
+                <tr>
+                    <th style="background-color:white;" style="text-align:center;">Title</th>
+                    <th style="background-color:white;" style="text-align:center;">Cover Photo</th>
+                    <th style="background-color:white;" style="text-align:center;">Channel</th>
+                </tr>
+            </thead>
+            <tbody id="results">
+            </tbody>
+        </table>
+        <div id="pager">
+            <ul id="pagination" class="pagination-sm"></ul>
+        </div>
+    </div>
+    <footer class ="footer">
+        <f>@drive sound</f>
+    </footer>
+</body>
+
+
+
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/2.4.4/umd/popper.min.js"></script>
+<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js"></script>
+<script src="jquery.twbsPagination.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/js/bootstrap-datepicker.min.js"></script>
+<script src="script.js"></script>
+
+</html>
+
+
+
+
+
+
