+++ Drive sound/web project/StarterrPage.html
@@ -16,8 +16,7 @@
                     <p>Website documentary music <br><br>You can create your music history and review some song.</p>
                 </div>
                 <div class ="input">
-                    <a href="Login.html" class="button">Login</a>
-                    <a href="Login.html" class="button">Register</a>
+                    <a href="Login.html" class="button">Join us</a>
                 </div>
             </div>
