+++ Drive sound/web project/css/styleforresult.css
@@ -1,0 +1,201 @@
+* {
+    box-sizing: border-box;
+  }
+body {
+    margin: 0px;
+    padding: 0px;
+    background-color: #2c3440;
+    font-family: Sans-serif;
+    color: #f2f2f2;
+    height: 100%;
+  }
+
+.header ,.footer {
+    height: 85px;
+    background-color: #242b3a;
+}
+.nav{
+    height: 85px;
+    float: right;
+}
+.nav a {
+    float: left;
+    color: #f2f2f2;
+    text-align: center;
+    padding: 33px 30px;
+    text-decoration: none;
+    font-size: 18px;
+  }
+  .nav a:hover {
+    background-color: #575c66;
+  }
+  .container{
+      margin-left: 25%;
+      margin-right: 25%;
+      margin-top: 1%;
+      height:748px;
+  }
+  
+  .column-left {
+    float: left;
+    width: 25%;
+    padding: 0 5px;
+  }
+  .column-mid {
+    float: left;
+    width: 50%;
+    padding: 0 5px;
+  }
+  .column-right {
+    float: left;
+    width: 25%;
+    padding: 0 5px;
+  }
+  .img-cover{
+      width: 180px;
+      height: 180px;
+  }
+  .clear {
+    clear: both;
+}
+.container a{
+    text-decoration: none;
+    color: inherit;
+}
+.container a:hover{
+    text-decoration: none;
+    color:aquamarine;
+}
+.container nav{
+    height: 52px;
+    background-color: #242b3a;
+    text-align: center;
+}
+.container nav a{
+    background-color: #242b3a;
+    padding-left: 5%;
+    padding-right: 5%;
+    padding-top: 4%;
+    padding-bottom: 4%;
+}
+.container nav a:hover{
+    background-color: #575c66;
+}
+.container nav a:active {
+    background-color: #575c66;
+  }
+.column-right li a{
+    display: block;
+    text-align: center;
+}
+.column-right .card{
+    margin-top: 40%;
+    list-style-type: none;
+}
+.next-page{
+    margin-top: 20%;
+    width: 100%;
+    justify-items: center;
+}
+
+
+
+/* CSS */
+.next-page a {
+  appearance: none;
+  background-color: #575c66;
+  border-width: 0;
+  box-sizing: border-box;
+  color: #ffffff;
+  cursor: pointer;
+  display: inline-block;
+  font-family: Clarkson,Helvetica,sans-serif;
+  font-size: 14px;
+  font-weight: 500;
+  letter-spacing: 0;
+  line-height: 1em;
+  margin: 0;
+  opacity: 1;
+  outline: 0;
+  padding: 1.5em 2.2em;
+  position: relative;
+  text-align: center;
+  text-decoration: none;
+  text-rendering: geometricprecision;
+  text-transform: uppercase;
+  transition: opacity 300ms cubic-bezier(.694, 0, 0.335, 1),background-color 100ms cubic-bezier(.694, 0, 0.335, 1),color 100ms cubic-bezier(.694, 0, 0.335, 1);
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+  vertical-align: baseline;
+  white-space: nowrap;
+  margin-left: 20%;
+}
+
+.next-page a:before {
+  animation: opacityFallbackOut .5s step-end forwards;
+  backface-visibility: hidden;
+  background-color: #242b3a;
+  clip-path: polygon(-1% 0, 0 0, -25% 100%, -1% 100%);
+  content: "";
+  height: 100%;
+  left: 0;
+  position: absolute;
+  top: 0;
+  transform: translateZ(0);
+  transition: clip-path .5s cubic-bezier(.165, 0.84, 0.44, 1), -webkit-clip-path .5s cubic-bezier(.165, 0.84, 0.44, 1);
+  width: 100%;
+}
+
+.next-page a:hover:before {
+  animation: opacityFallbackIn 0s step-start forwards;
+  clip-path: polygon(0 0, 101% 0, 101% 101%, 0 101%);
+}
+
+.next-page a:after {
+  background-color: #242b3a;
+}
+
+.next-page a span {
+  z-index: 1;
+  position: relative;
+}
+
+/* .header1 {
+    width: 1000px;
+    padding: 20px;
+    background: #5d6475;
+    margin: 25px auto;
+    border-radius: 5px;
+    text-align: center;
+}
+
+.header1 p {
+    font-size: 15px;
+    font-weight: 200;
+    color: #fff;
+}
+
+#i {
+    font-size: 45px;
+    text-transform: uppercase;
+    font-weight: 700;
+    color: #fff;
+    background: #2c3440;
+}
+
+body {
+    background: #2c3440;
+}
+
+body h1 {
+    background-color: #dceaff;
+}
+
+#im {
+    float: left;
+    padding: 14px 16px;
+    padding-top: 5px;
+    padding-bottom: 5px;
+    text-align: center;
+} */
