+++ Drive sound/web project/css/styleforstarter.css
@@ -38,13 +38,13 @@ body{
     font-size: larger;
   }
   .input{
-    margin-left: 17%;
+    margin-left: 20%;
     margin-top: 2%;
   }
  
 
 
-/* CSS */
+
 .button {
   align-items: center;
   background-color: #FFFFFF;
